+++ Adafruit_MQTT_Client.h
@@ -51,7 +51,7 @@ class Adafruit_MQTT_Client : public Adafruit_MQTT {
   bool disconnectServer();
   bool connected();
   uint16_t readPacket(uint8_t *buffer, uint16_t maxlen, int16_t timeout);
-  bool sendPacket(uint8_t *buffer, uint8_t len);
+  bool sendPacket(uint8_t *buffer, uint16_t len);
 
  private:
   Client* client;
